% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_01_7_PlotPSI_A5SS_PositiveStrand.R
\name{PlotPSI.A5SS.Pos}
\alias{PlotPSI.A5SS.Pos}
\title{Plot percent spliced-in (PSI) values for alternative 5' splice site (A5SS) located on positive strand}
\usage{
PlotPSI.A5SS.Pos(
  tran_id,
  Bam,
  BamPheno,
  cell.types,
  min.coverage,
  cons.exon.cutoff,
  method,
  method.adj,
  sig.pval = 0.1,
  cell.types.colors,
  plot.title,
  plot.width,
  plot.height,
  plot.out,
  track = TRUE,
  nboots = 2000,
  show.mean.ci = TRUE
)
}
\arguments{
\item{tran_id}{Character string. Splicing event nomenclature.}

\item{Bam}{Character string. Path to folder where the BAM files and their corresponding index files are located.}

\item{BamPheno}{object of class data.frame. Mandatory columns are \code{bam.file.name} and \code{cell.type}. \code{bam.file.name} column indicates BAM file names as per that found in the \code{Bam} folder. \code{cell.type} column indicates the cell group names.}

\item{cell.types}{Character string. Cell types to plot. Should be the same number of cell groups or less than the \code{cell.type} column of the \code{BamPheno} argument.}

\item{min.coverage}{Numeric value. Coverage (Total reads) threshold below which the PSI value of the genomic coordinate is annotate as missing value, i.e. no coverage.}

\item{cons.exon.cutoff}{Numeric value. Limit the number of bases to plot for the constitutive exons. This allow users to focus the plots on the alternative exon.}

\item{method}{Character string. Statistical test to compare the PSI values across the different cell types. \code{"wilcox"}, \code{"t.test"}, \code{"ks"}, \code{"ad"}, and \code{"dts"} available for 2-group comparison. \code{"ANOVA"} and \code{"kw"} available for 3- or more group comparison. \code{"ks"}, \code{"ad"}, \code{"dts"}, and \code{"kw"}, represent Kolmogorov–Smirnov, Anderson-Darling, DTS, and Kruskal-Wallis test, respectively.}

\item{method.adj}{Character string. Adjust p-values for multiple testing. Options available as per \code{p.adjust} function.}

\item{sig.pval}{Numeric value. Adjust p-value, below which, the p-value is considered statistically significant.}

\item{cell.types.colors}{Character string. Legend colors for each cell type. Should be of same length as \code{cell.types} argument. To use ggplot2 default color scheme, please specify \code{"ggplot.default"}.}

\item{plot.title}{Character string. Main title for plot. Examples are gene ID, gene names, splicing ID etc..}

\item{plot.width}{Numeric value. Width of plot.}

\item{plot.height}{Numeric value. Height of plot.}

\item{plot.out}{Character string. Path to folder to output plot.}

\item{track}{Logical. If set to \code{TRUE} (default), a process of reading in the BAM files, which is the rate-limiting step, will be tracked on the console.}

\item{nboots}{Numeric value. When \code{method} set to \code{"dts"}, the number of bootstrap iterations for computing the p-value.}

\item{show.mean.ci}{Logical value. If set to \code{TRUE}, the 95percent confidence interval of the per-cell group mean PSI values will not be shown. Default is \code{FALSE}.}
}
\value{
A plot in PDF format located in the folder specified by \code{plot.out} argument.
}
\description{
\code{PlotPSI.A5SS.Pos} computes percent spliced-in (PSI) at each genomic coordinate for alternative 5' splice site (A5SS) located on positive (forward) strand.
}
\details{
This function computes percent spliced-in (PSI) at each genomic coordinate for alternative 5' splice site (A5SS) located on positive (forward) strand. Formula for computing PSI is number of reads with non-N CIGAR operation divided by the total number of reads. Total number of reads is the sum of reads with non-N CIGAR operation and reads with N-CIGAR operation
}
\examples{
\donttest{
# DO NOT RUN
PlotPSI.A5SS.Pos(
 tran_id="chr3:197950190:197950221|197950299
          :+@chr3:197950936:197950978",
 Bam="/Users/BAM/",
 BamPheno=BamPheno,
 cell.types=c("iPSC", "Endoderm"),
 min.coverage=10,
 cons.exon.cutoff=100,
 method="ks",
 method.adj="bonferroni",
 cell.types.colors="ggplot.default",
 plot.title="SNRPN",
 plot.width=5,
 plot.height=8,
 plot.out=paste(tempdir(), "Plot.pdf", sep="")
 )
 }
}
\author{
Sean Wen <sean.wenwx@gmail.com>
}
