% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Script_01_0_PlotPSI.R
\name{PlotPSI}
\alias{PlotPSI}
\title{Plot percent spliced-in (PSI) values for exon-level alternative splicing events}
\usage{
PlotPSI(tran_id, event.type, strand, Bam, BamPheno, cell.types,
  min.coverage, cons.exon.cutoff, method, method.adj, cell.types.colors,
  plot.title, plot.width, plot.height, plot.out)
}
\arguments{
\item{tran_id}{Character string. Splicing event nomenclature.}

\item{event.type}{Character string. Specify \code{"SE"}, \code{"MXE"}, \code{"RI"}, \code{"A5SS"} or \code{"A3SS"}.}

\item{strand}{Character string. Specify \code{"positive"} or \code{"negative"} to indicate forward or negative strand, respectively.}

\item{Bam}{Character string. Path to folder where the BAM files and their corresponding index files are located.}

\item{BamPheno}{object of class data.frame. Mandatory columns are \code{bam.file.name} and \code{cell.type}. \code{bam.file.name} column indicates BAM file names as per that found in the \code{Bam} folder. \code{cell.type} column indicates the cell group names.}

\item{cell.types}{Character string. Cell types to plot. Should be the same number of cell groups or less than the \code{cell.type} column of the \code{BamPheno} argument.}

\item{min.coverage}{Numeric value. Coverage (Total reads) threshold below which the PSI value of the genomic coordinate is annotate as missing value, i.e. no coverage.}

\item{cons.exon.cutoff}{Numeric value. Limit the number of bases to plot for the constitutive exons. This allow users to focus the plots on the alternative exon.}

\item{method}{Character string. Statistical test to compare the PSI values across the different cell types. \code{"wilcox"}, \code{"t.test"}, and \code{"ks"} available for 2-group comparison. \code{"ANOVA"} and \code{"kw"} available for 3- or more group comparison. \code{"ks"} and \code{"kw"} represent Kolmogorov–Smirnov test and Kruskal-Wallis test, respectively.}

\item{method.adj}{Character string. Adjust p-values for multiple testing. Options available as per \code{p.adjust} function.}

\item{cell.types.colors}{Character string. Legend colors for each cell type. Should be of same length as \code{cell.types} argument. To use ggplot2 default color scheme, please specify \code{"ggplot.default"}.}

\item{plot.title}{Character string. Main title for plot. Examples are gene ID, gene names, splicing ID etc..}

\item{plot.width}{Numeric value. Width of plot.}

\item{plot.height}{Numeric value. Height of plot.}

\item{plot.out}{Character string. Path to folder to output plot.}
}
\value{
A plot in PDF format located in the folder specified by \code{plot.out} argument.
}
\description{
\code{PlotPSI} computes percent spliced-in (PSI) at each genomic coordinate for exon-level alternative splicing events.
}
\details{
This function computes percent spliced-in (PSI) at each genomic coordinate for exon-level alternative splicing events, namely skipped-exon (SE), mutually-exclusive exons (MXE), retained-intron (RI), alternative 5' splice site (A5SS), and alternative 3' splice site (A3SS). Formula for computing PSI is number of reads with non-N CIGAR operation divided by the total number of reads. Total number of reads is the sum of reads with non-N CIGAR operation and reads with N-CIGAR operation
}
\examples{
# Read sample metadata
path_to_file <- system.file("extdata", "BAM_PhenoData_Small.txt", package="VALERIE")
BamPheno <- read.table(path_to_file, sep="\\t", header=TRUE, stringsAsFactors=FALSE)
head(BamPheno)

# Plot
PlotPSI(
 tran_id="chr18:82554580:82554750:+@chr18:82561778:82561855:+@chr18:82572825:82572926",
 event.type="SE",
 strand="positive",
 Bam=system.file("extdata/BAM", package="VALERIE"),
 BamPheno=BamPheno,
 cell.types=c("Ctrl", "EAE"),
 min.coverage=10,
 cons.exon.cutoff=100,
 method="t.test",
 method.adj="bonferroni",
 cell.types.colors="ggplot.default",
 plot.title="Mbp",
 plot.width=5,
 plot.height=8,
 plot.out=paste(tempdir(), "Plot.pdf", sep="")
 )
}
\author{
Sean Wen <sean.wenwx@gmail.com>
}
